Change:split the file
