+++ 07.bowling_object/frame.rb
@@ -1,0 +1,28 @@
+# frozen_string_literal: true
+
+require './shot'
+
+class Frame
+  STRIKE = 10
+  attr_reader :first_shot, :second_shot, :third_shot
+
+  def initialize(first_shot, second_shot = nil, third_shot = nil)
+    @first_shot = Shot.new(first_shot)
+    @second_shot = Shot.new(second_shot)
+    @third_shot = Shot.new(third_shot)
+  end
+
+  def strike?
+    @first_shot.score == STRIKE
+  end
+
+  def spare?
+    @first_shot.score != STRIKE && [@first_shot.score, @second_shot.score].sum == 10
+  end
+
+  class << self
+    def pinfalls(pinfall_text)
+      pinfall_text.split(',').map { |pinfall| Shot.new(pinfall).score }
+    end
+  end
+end
