+++ 07.bowling_object/main.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+require './game'
+
+puts Game.new(ARGV[0]).calculate_score
