+++ 07.bowling_object/shot.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+class Shot
+  def initialize(pinfall)
+    @pinfall = pinfall
+  end
+
+  def score
+    @pinfall == 'X' ? 10 : @pinfall.to_i
+  end
+end
